Fixed build errors in mingw

+++ include/sockpp/sock_address.h
@@ -129,7 +129,7 @@ public:
 	 *  		  an address.
 	 */
 	sock_address_any(const sockaddr* addr, socklen_t n) {
-		if (n > MAX_SZ)
+		if (size_t(n) > MAX_SZ)
 			throw std::length_error("Address length out of range");
         std::memcpy(&addr_, addr, sz_ = n);
     }
@@ -141,7 +141,7 @@ public:
 	 *  		  an address.
 	 */
 	sock_address_any(const sockaddr_storage& addr, socklen_t n) {
-		if (n > MAX_SZ)
+		if (size_t(n) > MAX_SZ)
 			throw std::length_error("Address length out of range");
         std::memcpy(&addr_, &addr, sz_ = n);
     }
